Fixes from PVS-Studio static analysis
These changes could cause small behavior change, although they appear to be obvioiusly what was intended by the authors

[CL 2486879 by Dan Oconnor in Main branch]

+++ Engine/Source/Editor/SoundCueEditor/Private/SoundCueEditorUtilities.cpp
@@ -32,7 +32,7 @@ void FSoundCueEditorUtilities::CreateWaveContainers(TArray<USoundWave*>& Selecte
 {
 	const int32 NodeSpacing = 70;
 	
-	Location.Y -= (SelectedWaves.Num() - 1) * NodeSpacing / 2;
+	Location.Y -= static_cast<float>((SelectedWaves.Num() - 1) * NodeSpacing) / 2.f;
 
 	for(int32 WaveIndex = 0; WaveIndex < SelectedWaves.Num(); WaveIndex++)
 	{
